requirements.txt
uvicorn==0.13.2
fastapi==0.63.0
Pillow==7.1.2
paho-mqtt==1.5.1
aiofiles==0.6.0